+++ scripts/run_cibuildwheel.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-PYTHON_VERSION="3.10"
+PYTHON_VERSION="3.13"
 
 function usage {
   echo ""
@@ -53,7 +53,7 @@ VCPKG_DEFAULT_BINARY_CACHE=/host$VCPKG_HOST_BINARY_CACHE \
 VCPKG_INSTALLED_DIR=$VCPKG_INSTALLED_DIR \
 VCPKG_TARGET_TRIPLET=$VCPKG_TARGET_TRIPLET \
 LD_LIBRARY_PATH=$VCPKG_INSTALLED_DIR/$VCPKG_TARGET_TRIPLET/lib"
-export CIBW_REPAIR_WHEEL_COMMAND_LINUX="auditwheel repair -w {dest_dir} {wheel} --exclude libarrow_python.so.2000 --exclude libarrow.so.2000"
+export CIBW_REPAIR_WHEEL_COMMAND_LINUX="auditwheel repair -w {dest_dir} {wheel} --exclude libarrow_python.so.2100 --exclude libarrow.so.2100"
 # export CIBW_DEBUG_KEEP_CONTAINER=TRUE
 # export CIBW_DEBUG_TRACEBACK=TRUE
 # Use the following commands to inspect the stopped container: